pyproject.toml
[tool.poetry]
name = "ipca"
version = "0.1.0"
description = ""
authors = ["Thiago Oliveira <thiceconelo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.0"
loguru = "^0.6.0"
sidrapy = "^0.1.3"
openpyxl = "^3.0.10"
python-crontab = "^2.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
